+++ library/src/methods/parse/parse.ts
@@ -7,6 +7,7 @@ import type {
   InferOutput,
 } from '../../types/index.ts';
 import { ValiError } from '../../utils/index.ts';
+import type { RecurResolved } from '../recursive/types.ts';
 
 /**
  * Parses an unknown input based on a schema.
@@ -20,7 +21,7 @@ import { ValiError } from '../../utils/index.ts';
 export function parse<
   const TSchema extends BaseSchema<unknown, unknown, BaseIssue<unknown>>,
 >(
-  schema: TSchema,
+  schema: TSchema & RecurResolved<TSchema>,
   input: unknown,
   config?: Config<InferIssue<TSchema>>
 ): InferOutput<TSchema> {